refactored actionTypes folder
