+++ .TODO
@@ -4,8 +4,6 @@
 * Типизировать setCookie.ts
 * Оформить NotFoundPage
 * Перенести fetch в redux из ForgotPasswordPage
-* Мемоизировать новые pages
-* В RegisterPage сделать автологин после регистрации
 * Сделать проверки на наличие токенов в куки и лс, чтобы не дублировать их и не вызвать зацикленность
 * Убрать лишние useMemo и useCallback для наилучшей оптимизации
 * Деструктурировать компонент BurgerConstructor. Слишком много строк для одного компонента
