+++ src/redux/actionTypes/orderTypes.ts
@@ -1,0 +1,14 @@
+export interface TOrderState {
+  orderInfo: {
+    name: string
+    order: number
+  }
+  isLoading: boolean
+  hasError: boolean
+}
+
+export enum OrderFetchStatus {
+  ORDER_REQUEST = 'ORDER_REQUEST',
+  ORDER_SUCCESS = 'ORDER_SUCCESS',
+  ORDER_ERROR = 'ORDER_ERROR',
+}
